refactor test cases for relocated profile import
